# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Fiber API with OAuth2 password flow",
        "title": "Openapi spec",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/car/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieves a list of cars with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Retrieves lists cars with optional search, type, and pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CarPaginatedResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/car/count/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieves a count of cars",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Count returns total cars count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "count: 12345",
                        "schema": {
                            "$ref": "#/definitions/CountResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/car/create/": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Creates new car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Creates a new car",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CarCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/CarMessageResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "422": {
                        "description": "detail: Validation errors",
                        "schema": {
                            "$ref": "#/definitions/ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/car/{id}/detail/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieve single car detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Retrieve single car",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Car ID (int)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CarResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/car/{id}/update/": {
            "patch": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Updates an existing car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Update a car",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Car ID (int)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request data",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CarUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CarMessageResponse"
                        }
                    },
                    "400": {
                        "description": "detail: BadRequest - invalid request",
                        "schema": {
                            "$ref": "#/definitions/BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "404": {
                        "description": "detail: User not found",
                        "schema": {
                            "$ref": "#/definitions/NotFoundResponse"
                        }
                    },
                    "422": {
                        "description": "detail: Validation errors",
                        "schema": {
                            "$ref": "#/definitions/ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "BadRequestResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "BadRequest - invalid request"
                }
            }
        },
...
        "CarCreateInput": {
            "type": "object",
            "required": [
                "carNumber"
            ],
            "properties": {
                "carNumber": {
                    "type": "string",
                    "maxLength": 255
                },
                "isStaff": {
                    "type": "boolean"
                },
                "ownerName": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "CarMessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/CarResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "CarPaginatedResponse": {
            "type": "object",
            "required": [
                "limit",
                "page",
                "rows"
            ],
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CarResponse"
                    }
                }
            }
        },
...
        "CarResponse": {
            "type": "object",
            "required": [
                "carNumber",
                "createdAt",
                "id",
                "isStaff"
            ],
            "properties": {
                "carNumber": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isStaff": {
                    "type": "boolean"
                },
                "ownerName": {
                    "type": "string"
                },
                "updatedAt": {
                    "description": "optional, can be nil",
                    "type": "string"
                }
            }
        },
...
        "CarUpdateInput": {
            "type": "object",
            "properties": {
                "carNumber": {
                    "type": "string",
                    "maxLength": 255
                },
                "isStaff": {
                    "type": "boolean"
                },
                "ownerName": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "CountResponse": {
            "type": "object",
            "required": [
                "count"
            ],
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 12345
                }
            }
        },
...
        "InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
        "NotFoundResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Not found"
                }
            }
        },
        "PermissionDeniedResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Permission denied"
                }
            }
        },
...
        "UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Unauthorized - Invalid token"
                }
            }
        },
...
        "ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Validation error"
                },
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "fieldName": "this field is required"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/v1/auth/login",
            "scopes": {
                "read": "Grants read access",
                "write": "Grants write access"
            }
        }
    }
}